+++ src/types/stats.rs
@@ -7,5 +7,5 @@ use serde::{Deserialize, Serialize};
 pub struct Stats {
     pub average_median_fees: HashMap<u32, f64>,
     pub transactions: HashMap<u32, u64>,
-    pub volume: HashMap<u32, f64>,
+    pub volume: HashMap<u32, u64>,
 }
